Add files via upload
To use XAI, start by running the XAI-Shear.py script. Since the code originated from R and is designed to run in a Google Colab environment, you will need to uncomment the necessary code and open it in Google Colab. Use the dataset named M1Cylindrical.xlsx to load the data, and utilize the script "gp-main-for-eq-that-reported" to generate an equation based on the data from the Excel file. For verification purposes, refer to the dataset labeled "data-for-verification". To cite the paper, please use this: